Removed Item and User classes, replaced by MyItem and MyUser + solved issue in sign-in tests
