+++ app/src/androidTest/java/com/github/brugapp/brug/sign_in/SignInActivityTestFakeHandler.kt
@@ -176,6 +176,7 @@ class SignInActivityTestFakeHandler {
 
 
         ActivityScenario.launch<SignInActivity>(intent).use {
+            Thread.sleep(1000)
             onView(withId(R.id.sign_in_google_button)).perform(click())
         }
     }
