+++ app/src/main/java/com/github/brugapp/brug/model/Conversation.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 
 data class Conversation(
     val convId: String,
-    val userFields: MyUser,
+    val userFields: User,
     val lostItemName: String,
     val lastMessage: Message?
 ) : Serializable
