+++ app/src/main/java/com/github/brugapp/brug/model/Item.kt
@@ -1,76 +1,87 @@
 package com.github.brugapp.brug.model
 
 import com.github.brugapp.brug.R
+import com.github.brugapp.brug.model.services.LocationService
 import java.io.Serializable
 
-class Item// Generate the id later
-    (private var name: String, private var description: String, private var id: String) :
-    Serializable {
+class Item(
+    val itemName: String,
+    val itemTypeID: Int,
+    val itemDesc: String,
+    private var isLost: Boolean
+) : Serializable {
 
-    // add QR code attribute
-    // add last localization attribute
-    // image_id attribute is for future use (with image type)
+    /* ITEM ID */
+    private var itemID: String = ""
+    private var lastLocation: LocationService? = null
 
-    private var lost: Boolean = false
-    private var type: ItemType = ItemType.Other
-    private var iconId: Int = 0
-
-    init {
-        if (name.isBlank()) {
-            throw IllegalArgumentException("Invalid name")
-        }
+    fun getLastLocation(): LocationService? {
+        return lastLocation
+    }
+    fun setLastLocation(lon: Double, lat: Double){
+        lastLocation = LocationService(lat, lon)
     }
 
+    fun setItemID(itemID: String) {
+        this.itemID = itemID
+    }
+
+    fun getItemID(): String {
+        return this.itemID
+    }
+
+    /* ITEM TYPE */
+    fun getRelatedIcon(): Int {
+        return getRelatedItemType().getRelatedIcon()
+    }
+
+    private fun getRelatedItemType(): ItemType {
+        return if (0 <= itemTypeID && itemTypeID < ItemType.values().size)
+            ItemType.values()[itemTypeID]
+        else ItemType.Other
+    }
+
+    /* ITEM STATE */
     fun isLost(): Boolean {
-        return lost
+        return this.isLost
     }
 
-    fun setLost(value: Boolean): Item {
-        this.lost = value
-        return this
+    fun changeLostStatus(newStatus: Boolean) {
+        this.isLost = newStatus
     }
 
-    fun getName(): String {
-        return name
+    override fun equals(other: Any?): Boolean {
+        val otherItem = other as Item
+        return this.itemID == otherItem.itemID
+                && this.itemName == otherItem.itemName
+                && this.itemTypeID == otherItem.itemTypeID
+                && this.itemDesc == otherItem.itemDesc
+                && this.isLost == otherItem.isLost
     }
 
-
-    fun getId(): String {
-        return id
+    override fun hashCode(): Int {
+        var result = itemID.hashCode()
+        result = 31 * result + itemName.hashCode()
+        result = 31 * result + itemTypeID
+        result = 31 * result + itemDesc.hashCode()
+        result = 31 * result + isLost.hashCode()
+        return result
     }
 
-    fun getDescription(): String {
-        return description
-    }
+}
 
-    fun setName(name: String): Item {
+enum class ItemType {
+    Wallet, Keys,CarKeys,Phone,Other;
 
-        if (name.isBlank()) {
-            throw IllegalArgumentException("Invalid new name")
-        }
-
-        this.name = name
-        return this
-    }
-
-    fun setType(type: ItemType): Item {
-        this.type = type
-        iconId = when (type) {
+    /* ITEM TYPE */
+    fun getRelatedIcon(): Int {
+        return when(this){
             ItemType.Wallet -> R.drawable.ic_baseline_account_balance_wallet_24
             ItemType.Keys -> R.drawable.ic_baseline_vpn_key_24
             ItemType.CarKeys -> R.drawable.ic_baseline_car_rental_24
             ItemType.Phone -> R.drawable.ic_baseline_smartphone_24
             ItemType.Other -> R.drawable.ic_baseline_add_24
+            else -> R.drawable.ic_baseline_add_24
         }
-        return this
-    }
-
-    fun getIcon(): Int {
-        return iconId
-    }
-
-    fun setDescription(description: String) {
-        this.description = description
     }
 }
-
