+++ app/src/main/java/com/github/brugapp/brug/ui/ItemsMenuActivity.kt
@@ -16,10 +16,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.github.brugapp.brug.ITEMS_TEST_LIST_KEY
 import com.github.brugapp.brug.ITEM_INTENT_KEY
-import com.github.brugapp.brug.MESSAGE_TEST_LIST_KEY
 import com.github.brugapp.brug.R
 import com.github.brugapp.brug.data.ItemsRepository
-import com.github.brugapp.brug.model.MyItem
+import com.github.brugapp.brug.model.Item
 import com.github.brugapp.brug.ui.components.BottomNavBar
 import com.github.brugapp.brug.ui.components.CustomTopBar
 import com.github.brugapp.brug.view_model.ItemsListAdapter
@@ -70,11 +69,11 @@ class ItemsMenuActivity : AppCompatActivity() {
 
     // ONLY GETS THE LIST OF ITEMS RELATED TO THE USER, NOT THE FULL USER PROFILE !
     private fun initItemsList() {
-        val observableList = MutableLiveData<MutableList<MyItem>>()
+        val observableList = MutableLiveData<MutableList<Item>>()
 
         val itemsTestList =
             if(intent.extras != null && intent.extras!!.containsKey(ITEMS_TEST_LIST_KEY)){
-                intent.extras!!.get(ITEMS_TEST_LIST_KEY) as MutableList<MyItem>
+                intent.extras!!.get(ITEMS_TEST_LIST_KEY) as MutableList<Item>
             } else null
 
         viewModel.viewModelScope.launch {
