+++ app/src/main/java/com/github/brugapp/brug/view_model/ItemInformationViewModel.kt
@@ -1,15 +1,15 @@
 package com.github.brugapp.brug.view_model
 
 import androidx.lifecycle.ViewModel
-import com.github.brugapp.brug.model.MyItem
+import com.github.brugapp.brug.model.Item
 import com.google.firebase.auth.FirebaseAuth
 
 class ItemInformationViewModel : ViewModel() {
 
     private lateinit var qrId: String
-    private lateinit var item: MyItem
+    private lateinit var item: Item
 
-    fun getText(item: MyItem, firebaseAuth: FirebaseAuth): HashMap<String, String> {
+    fun getText(item: Item, firebaseAuth: FirebaseAuth): HashMap<String, String> {
         this.item = item
         qrId = firebaseAuth.uid + ":" + item.getItemID()
         val hash: HashMap<String, String> = HashMap()
