+++ app/src/test/java/com/github/brugapp/brug/UserTest.kt
@@ -1,100 +1,38 @@
 package com.github.brugapp.brug
 
-import com.github.brugapp.brug.model.Item
 import com.github.brugapp.brug.model.User
-import org.hamcrest.MatcherAssert.*
-import org.junit.Assert
-import java.lang.IllegalArgumentException
+import org.hamcrest.MatcherAssert.assertThat
+import org.hamcrest.core.IsEqual
+import org.hamcrest.core.IsNot
+import org.hamcrest.core.IsNull
 import org.junit.Test
-import org.hamcrest.CoreMatchers.`is` as Is
-
 
 class UserTest {
-
     @Test
-    fun invalidFirstName() {
-        val firstName = "   "
-        Assert.assertThrows(IllegalArgumentException::class.java) {
-            User(firstName, "Kikou", "rayan.kikou@gmail.com", "0",null)
-        }
-    }
-
-    @Test
-    fun emptyFirstName() {
-        val firstName = ""
-        Assert.assertThrows(IllegalArgumentException::class.java) {
-            User(firstName, "Kikou", "rayan.kikou@gmail.com", "0",null)
-        }
-    }
-
-    @Test
-    fun invalidLastName() {
-        val lastName = "   "
-        Assert.assertThrows(IllegalArgumentException::class.java) {
-            User("Rayan", lastName, "rayan.kikou@gmail.com", "0",null)
-        }
-    }
-
-    @Test
-    fun emptyLastName() {
-        val lastName = ""
-        Assert.assertThrows(IllegalArgumentException::class.java) {
-            User("Rayan", lastName, "rayan.kikou@gmail.com", "0",null)
-        }
-    }
-
-    @Test
-    fun invalidEmail() {
-        val email = "test.com"
-        Assert.assertThrows(IllegalArgumentException::class.java) {
-            User("Rayan", "Kikou", email, "0",null)
-        }
-
-    }
-
-    @Test
-    fun invalidID() {
-        val id = ""
-        Assert.assertThrows(IllegalArgumentException::class.java) {
-            User("Rayan", "Kikou", "rayan.kikou@gmail.com", id,null)
-        }
-
-    }
-
-    @Test
-    fun validUser(){
-
+    fun initUserWithoutIconCorrectlyInitializesUser() {
+        val userID = "DUMMYID"
         val firstName = "Rayan"
         val lastName = "Kikou"
-        val email = "rayan.kikou@gmail.com"
-        val id = "0"
+        val userIconPath: String? = null
 
-        val validUser = User(firstName, lastName, email, id, null)
-
-        assertThat(1, Is(1))
-        assertThat(validUser.getFirstName(), Is(firstName))
-        assertThat(validUser.getLastName(), Is(lastName))
-        assertThat(validUser.getEmail(), Is(email))
-        assertThat(validUser.getId(), Is(id))
+        val user = User(userID, firstName, lastName, userIconPath, mutableListOf())
+        assertThat(user.uid, IsEqual(userID))
+        assertThat(user.firstName, IsEqual(firstName))
+        assertThat(user.lastName, IsEqual(lastName))
+        assertThat(user.getUserIconPath(), IsNull.nullValue())
     }
 
     @Test
-    fun getUserItemListTest(){
-
-        val firstName = "Rayan"
-        val lastName = "Kikou"
-        val email = "rayan.kikou@gmail.com"
-        val id = "0"
-
-        val validUser = User(firstName, lastName, email, id, null)
-
-        val newItem = Item("Wallet", "Grey wallet", "0")
-        var itemList = ArrayList<Item>(validUser.getItemList())
-        validUser.addItem(newItem)
-        itemList.add(newItem)
-
-        assertThat(validUser.getItemList(), Is(itemList))
+    fun comparingTwoIdenticalUsersReturnsEquality() {
+        val user1 = User("DUMMYID", "Rayan", "Kikou", null, mutableListOf())
+        val user2 = User("DUMMYID", "Rayan", "Kikou", null, mutableListOf())
+        assertThat(user1, IsEqual(user2))
     }
 
-
+    @Test
+    fun comparingTwoAlmostIdenticalUsersReturnsFalse() {
+        val user1 = User("DUMMYID", "Rayan", "Kikou", null, mutableListOf())
+        val user2 = User("DUMMYID2", "Rayan", "Kikou", null, mutableListOf())
+        assertThat(user1, IsNot(IsEqual(user2)))
+    }
 }
